Lab8

+++ Lab7/src/routes/orders.py
@@ -3,6 +3,12 @@ from flask import Blueprint, jsonify, request
 from src.utils import *
 import src.models as models
 import src.db as db
+from src.routes.user import UserRole
+from src import app
+from flask_jwt_extended import (
+    JWTManager, jwt_required, get_jwt_identity
+)
+jwt = JWTManager(app)
 
 order_bp = Blueprint('order', __name__, url_prefix='/order')
 
@@ -14,6 +20,7 @@ def get_books():
 
 
 @order_bp.route('/store/make_order', methods=['POST'])
+@jwt_required()
 def create_order():
     class Order(Schema):
         ship_date = fields.Date(required=True)
@@ -26,6 +33,16 @@ def create_order():
         order = Order().load(request.json)
     except ValidationError as error:
         return jsonify(error.messages), 400
+
+    user = db.session.query(models.User).filter(models.User.id == get_jwt_identity()).first()
+    book = db.session.query(models.Book).filter(models.Book.id == request.json['book_id']).first()
+
+    if user is None:
+        return jsonify({'message': 'Unauthorized'}), 404
+    if user.id != request.json['user_id']:
+        return jsonify({'message': 'Wrong user_id'}), 404
+    if book is None:
+        return jsonify({'message': 'Book not found'}), 404
 
     new_order = models.Order(ship_date=order['ship_date'], complete=order['complete'],
                              address=order['address'], user_id=order['user_id'], book_id=order['book_id'])
@@ -49,6 +66,7 @@ def get_order_id(order_id):
 
 
 @order_bp.route('/store/<int:order_id>/', methods=['PUT'])
+@jwt_required()
 def update_order(order_id):
     class Order(Schema):
         new_ship_date = fields.Date(required=False)
@@ -61,6 +79,11 @@ def update_order(order_id):
         order = Order().load(request.json)
     except ValidationError as error:
         return jsonify(error.messages), 400
+    user = db.session.query(models.User).filter(models.User.id == get_jwt_identity()).first()
+    if user is None:
+        return jsonify({'message': 'Unauthorized'}), 404
+    if user.role != UserRole.Administrator:
+        return jsonify({'message': 'Forbidden'}), 401
     order_to_update = db.session.query(models.Order).filter(models.Order.id == order_id).first()
     if order_to_update is None:
         return jsonify({'message': 'Book not found'}), 404
@@ -83,10 +106,16 @@ def update_order(order_id):
 
 
 @order_bp.route('/store/<int:order_id>/', methods=['DELETE'])
+@jwt_required()
 def delete_order(order_id):
     order = db.session.query(models.Order).filter(models.Order.id == order_id).first()
     if order is None:
         return jsonify({'message': 'Order not found'}), 404
+    user = db.session.query(models.User).filter(models.User.id == get_jwt_identity()).first()
+    if user is None:
+        return jsonify({'message': 'Unauthorized'}), 404
+    if user.role != UserRole.Administrator:
+        return jsonify({'message': 'Forbidden'}), 401
     try:
         db.session.delete(order)
     except:
